clean up vid utils
