+++ multicamera_acquisition/tests/writer/test_writers.py
@@ -11,7 +11,7 @@ import pytest
 
 from multicamera_acquisition.writer import FFMPEG_Writer
 
-from multicamera_acquisition.video_io_ffmpeg import count_frames
+from multicamera_acquisition.video_utils import count_frames
 
 
 @pytest.fixture(scope="session")
